vfio: Add support for mapping mmap'd mmio pages
To support peer to peer DMA, we need to support more than
just get_user_pages_fast address lookups.  This patch
leverages a stripped down version of KVM's hva_to_pfn to
enable these kinds of translations.

Since this means that not every mapping in the iommu is
backed by a struct page, we can get rid of the wasted time
and space of maintaining them.  Instead we again take the
KVM approach of reading the physical translation of an
iova out of the iommu to unpin.  This greatly simplifies
much of the region shinking/splitting/merging code.

Note that KVM keeps a fault page for mapping areas that would
hit the BUG() in the patch below.  I'm not convinced we need
such a slush page for vfio, so I'm curious if we ever
legitimately hit the BUG().

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>