vfio: Increase PCI region address space
We currently only allow 32bits of address space for each PCI
region.  It's not too inconceivable to have a 4GB MMIO region
for a device supporting 64bit BARs.  Bump this up to 48bit
support, which still gives us 16bits to segment and flag the
regions, of which we're currently only using 4bits.  This
gives us 256TB per region.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>