vfio: Unmap ioctl returns region unmapped
Just like the map ioctl returns the region actually mapped, let
the unmap ioctl return the region unmapped.  We currently break
down mappings to individual pages, and can therefore unmap any
subset of a previous mapping, but that might not always be the
case.  In particular, we'll want to make use of large pages
supported by an iommu, which may not always be able to be split.
This gives us that flexibility.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>